+++ boofuzz/primitives/from_file.py
@@ -5,15 +5,13 @@ from .base_primitive import BasePrimitive
 
 
 class FromFile(BasePrimitive):
-    def __init__(self, value, encoding="ascii", fuzzable=True, max_len=0, name=None, filename=None):
+    def __init__(self, value, fuzzable=True, max_len=0, name=None, filename=None):
         """
         Cycles through a list of "bad" values from a file(s). Takes filename and open the file(s) to read
         the values to use in fuzzing process. filename may contain glob characters.
 
         @type  value:    str
         @param value:    Default string value
-        @type  encoding: str
-        @param encoding: (Optional, def="ascii") String encoding, ex: utf_16_le for Microsoft Unicode.
         @type  fuzzable: bool
         @param fuzzable: (Optional, def=True) Enable/disable fuzzing of this primitive
         @type  max_len:  int
@@ -27,14 +25,13 @@ class FromFile(BasePrimitive):
         super(FromFile, self).__init__()
 
         self._value = self._original_value = value
-        self.encoding = encoding
         self._fuzzable = fuzzable
         self._name = name
         self._filename = filename
         self._fuzz_library = []
         list_of_files = glob.glob(self._filename)
         for fname in list_of_files:
-            with open(fname, "r") as _file_handle:
+            with open(fname, "rb") as _file_handle:
                 self._fuzz_library.extend(_file_handle.readlines())
 
         # TODO: Make this more clear